requirements.txt
accelerate
torch
torchvision
numpy
opencv-python
pillow
transformers
scipy
diffusers